+++ libs/base/src/code/util/_reexport.rs
@@ -341,7 +341,6 @@ macro_rules! __reexport {
 }
 pub use __reexport as _reexport;
 
-
 /// A handy helper to include a rust source file keeping the intermediary docs.
 ///
 /// This is used to include files re-exported by lower-level workspace crates,